Conflicts:
	README.md

On branch master
All conflicts fixed but you are still merging.

 Changes to be committed:

	modified:   README.md
	modified:   configuresudoers.sh
	new file:   init.sh
	modified:   scripts/aliases.sh
	new file:   scripts/firewalld.sh
	modified:   scripts/lemp.sh
	modified:   scripts/linux_update.sh
	modified:   scripts/middleman.sh
	modified:   scripts/nginx_config.sh
	deleted:    scripts/server_firewall.sh
	modified:   scripts/server_ssh.sh
	modified:   scripts/swap.sh
	modified:   scripts/wordpress_install.sh
	deleted:    scripts/workstation_firewall.sh
	deleted:    scripts/workstation_ssh.sh
	modified:   setup.sh

+++ scripts/middleman.sh
@@ -1,8 +1,7 @@
 #!/bin/bash
 echo "*********************************************"
-echo "* A CentOS 7.0 deployment script to          "
-echo "* install Middleman and dependencies and     "
-echo "* deploy to BitBalloon                       "
+echo "* A CentOS 7.0 x64 deployment script to      "
+echo "* install Middleman and dependencies         "
 echo "*                                            "
 echo "* Author : Keegan Mullaney                   "
 echo "* Company: KM Authorized LLC                 "
@@ -19,9 +18,19 @@ else
    curl -L https://get.rvm.io | bash -s stable --ruby=$RUBY_VERSION
 fi
 
-# update gem
+# start using rvm
 echo
-read -p "Press enter to update gem..."
+read -p "Press enter to start using rvm..."
+if cat /home/$USER_NAME/.bashrc | grep -q "/usr/local/rvm/scripts/rvm"; then
+   echo "already added rvm to .bashrc"
+else
+   echo "source /usr/local/rvm/scripts/rvm" >> /home/$USER_NAME/.bashrc
+   source /usr/local/rvm/scripts/rvm && echo "rvm sourced and added to .bashrc"
+fi
+
+# update gems
+echo
+read -p "Press enter to update gems..."
 gem update
 
 echo
@@ -34,24 +43,31 @@ gem update --system
 #yum --enablerepo=epel -y install nodejs npm
 
 # install Middleman
-echo
-read -p "Press enter to install middleman..."
-gem install middleman
+if $(gem list middleman -i); then
+   echo "middleman gem already installed"
+else
+   echo
+   read -p "Press enter to install middleman..."
+   gem install middleman
+fi
 
 # install Redcarpet (for Markdown file processing)
-echo
-read -p "Press enter to install redcarpet..."
-gem install redcarpet
+if $(gem list redcarpet -i); then
+   echo "redcarpet gem already installed"
+else
+   echo
+   read -p "Press enter to install redcarpet..."
+   gem install redcarpet
+fi
 
 # install Rouge (for code syntax highlighting)
-echo
-read -p "Press enter to install rouge..."
-gem install rouge
-
-# install BitBalloon gem
-#echo
-#read -p "Press enter to install the BitBalloon gem..."
-#gem install bitballoon
+if $(gem list rouge -i); then
+   echo "rouge gem already installed"
+else
+   echo
+   read -p "Press enter to install rouge..."
+   gem install rouge
+fi
 
 # Middleman web root
 #mkdir -p /var/www/$MIDDLEMAN_DOMAIN/public_html
@@ -59,29 +75,27 @@ gem install rouge
 #echo "made directory: $_ and set permissions to $USER_NAME"
 
 # Middleman repository
-echo
-read -p "Press enter to create repos directory for $USER_NAME..."
 MM_DIRECTORY="/home/$USER_NAME/repos/$MIDDLEMAN_DOMAIN"
 if [ -d $MM_DIRECTORY ]; then
    echo "$MM_DIRECTORY directory already exists"
 else
+   echo
+   read -p "Press enter to create repos directory for $USER_NAME..."
    mkdir -p $MM_DIRECTORY
    echo "made directory: $_"
 fi
 
 # change to repos directory
-echo
-read -p "Press enter to change to repos directory..."
 cd $MM_DIRECTORY
-echo "changed directory to $_"
+echo "changing directory to $_"
 
 # generate a blog template for Middleman
-echo
-echo "Before proceeding, make sure to fork $UPSTREAM_REPO and change the project name to $MIDDLEMAN_PROJECT on GitHub"
-read -p "Press enter to clone $MIDDLEMAN_PROJECT from GitHub..."
 if [ -d "$MM_DIRECTORY/$MIDDLEMAN_PROJECT" ]; then
    echo "$MIDDLEMAN_PROJECT directory already exists, skipping clone operation..."
 else
+   echo
+   echo "Before proceeding, make sure to fork $UPSTREAM_REPO and change the project name to $MIDDLEMAN_PROJECT on GitHub"
+   read -p "Press enter to clone $MIDDLEMAN_PROJECT from GitHub..."
    echo
    echo "Do you wish to clone using HTTPS or SSH (recommended)?"
    select hs in "HTTPS" "SSH"; do
@@ -98,17 +112,17 @@ else
 fi
 
 # change to newly cloned directory
-echo
-read -p "Press enter to change to set permissions and cd to project directory..."
-chown -R $USER_NAME:$USER_NAME $MM_DIRECTORY
-echo "set permissions on $MM_DIRECTORY to $USER_NAME"
 cd $MIDDLEMAN_PROJECT
-echo "changed directory to $_"
+echo "changing directory to $_"
 
 # assign the original repository to a remote called "upstream"
-echo
-read -p "Press enter to assign upstream repository..."
-git remote add upstream https://github.com/$UPSTREAM_REPO && echo "remote upstream added for https://github.com/$UPSTREAM_REPO"
+if git config --list | grep -q $UPSTREAM_REPO; then
+   echo "upstream repo already configured: https://github.com/$UPSTREAM_REPO"
+else
+   echo
+   read -p "Press enter to assign upstream repository..."
+   git remote add upstream https://github.com/$UPSTREAM_REPO && echo "remote upstream added for https://github.com/$UPSTREAM_REPO"
+fi
 
 # pull in changes not present local repository, without modifying local files
 echo
@@ -122,19 +136,17 @@ read -p "Press enter to merge changes..."
 git merge upstream/master
 
 # specify middleman-bitballoon extension in the Gemfile
-echo
-read -p "Press enter to configure the Gemfile..."
-egrep -i "rouge" Gemfile
-if [ $? -eq 0 ]; then
+if cat Gemfile | grep -q "rouge"; then
    echo "Rouge syntax highligting already configured"
 else
+   echo
+   read -p "Press enter to configure the Gemfile..."
    echo '# Ruby based syntax highlighting' >> Gemfile
    echo 'gem "rouge"' >> Gemfile
    echo "rouge added to Gemfile"
-fi
-egrep -i "middleman-bitballoon" Gemfile
-if [ $? -eq 0 ]; then
-   echo "BitBalloon extension already configured"
+fi 
+if cat Gemfile | grep -q "middleman-bitballoon"; then
+   echo "BitBalloon extension already configured in Gemfile"
 else
    echo '' >> Gemfile
    echo '# Middleman extension for deploying to BitBalloon' >> Gemfile
@@ -143,17 +155,15 @@ else
 fi
 
 # configure BitBalloon extension in config.rb
-echo
-read -p "Press enter to configure config.rb..."
-egrep -i "bitballoon.build_before" config.rb
-if [ $? -eq 0 ]; then
-   echo "BitBalloon extension already configured"
+if cat config.rb | grep -q "bitballoon.build_before"; then
+   echo "BitBalloon extension already configured in config.rb"
 else
-   egrep -i "BB_TOKEN" /home/$USER_NAME/.bash_profile
-   if [ $? -eq 0 ]; then
+   echo
+   read -p "Press enter to configure config.rb..."
+   if cat /home/$USER_NAME/.bash_profile | grep -q "BB_TOKEN"; then
       echo "BB_TOKEN already entered in .bash_profile for user: $USER_NAME"
    else
-      read -e -p "Paste your BitBalloon app token here..." GET_TOKEN
+      read -e -p "Paste your BitBalloon app token here: " GET_TOKEN
       echo -e "\nexport BB_TOKEN=${GET_TOKEN}" >> /home/$USER_NAME/.bash_profile
    fi
    cat << EOF >> config.rb
@@ -169,9 +179,16 @@ EOF
    echo "BitBalloon extension configured"
 fi
 
-# change back to home directory
-cd /home/$USER_NAME
-echo "changed directory to /home/$USER_NAME"
+# set permissions
 echo
+read -p "Press enter to change to set permissions..."
+chown -R $USER_NAME:$USER_NAME $MM_DIRECTORY
+echo "set permissions on $MM_DIRECTORY to $USER_NAME"
+
+# update gems
+echo
+read -p "Press enter to update gems..."
+gem update
+
 echo "done with middleman.sh"
